+++ Frontend/Vue/src/views/Activities.vue
@@ -1,21 +1,29 @@
 <template>
     <div class="activities">
       <div class="container-fluid px-4">
-        <h1 class="mb-4 mt-4">Activities</h1>
+        <h1 class="mb-4 mt-4 text-center title">Activities</h1>
         <div class="row">
           <div class="col-xl-3 col-md-6 mb-4" v-for="activity in activities" :key="activity.name">
             <div class="card text-center bg-light text-dark h-100">
               <img :src="activity.image" class="card-img-top" :alt="activity.name">
               <div class="card-body">
-                <div class="card-body-icon">
-                  <i :class="activity.icon"></i>
-                </div>
                 <div class="fw-bold">{{ activity.name }}</div>
                 <router-link :to="{ name: 'LogActivity', params: { activityName: activity.name } }" class="btn btn-primary mt-3">
                   Log {{ activity.name }}
                 </router-link>
               </div>
             </div>
+          </div>
+          <div class="col-xl-3 col-md-6 mb-4">
+          <div class="card text-center bg-light text-dark h-100">
+              <img src='@/assets/images/custom.png' class="card-img-top" alt="custom">
+              <div class="card-body">
+                <div class="fw-bold">Add your own activity</div>
+                <router-link :to="{ name: 'LogActivity', params: { activityName: 'Activity' } }" class="btn btn-primary mt-3">
+                  Log Activity
+                </router-link>
+              </div>
+           </div>
           </div>
         </div>
       </div>
@@ -28,10 +36,11 @@
     data() {
       return {
         activities: [
-          { name: 'Running', image: 'src/assets/images/running.png', icon: 'fas fa-running' },
-          { name: 'Cycling', image: 'src/assets/images/cycling.png', icon: 'fas fa-biking' },
-          { name: 'Swimming', image: 'src/assets/images/swimming.png', icon: 'fas fa-swimmer' },
-          { name: 'Weightlifting', image: 'src/assets/images/weightlifting.png', icon: 'fas fa-dumbbell' }
+          { name: 'Running', image: 'src/assets/images/running.png' },
+          { name: 'Cycling', image: 'src/assets/images/cycling.png'},
+          { name: 'Swimming', image: 'src/assets/images/swimming.png' },
+          { name: 'Weightlifting', image: 'src/assets/images/weightlifting.png' },
+          { name: 'Walking', image: 'src/assets/images/walking.png'}
         ]
       };
     }
@@ -41,6 +50,9 @@
   <style scoped>
   .activities {
     margin-top: 20px;
+  }
+  .title {
+    color: #57B846;
   }
   .card {
     border: none;
@@ -68,6 +80,7 @@
     border: none;
     border-radius: 25px;
     font-family: 'Montserrat', sans-serif;
+    font-weight: 700;
     padding: 10px 20px;
   }
   .btn-primary:hover {