+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+
+## Jul 31, 2024
+- Improved UI of all components for better aesthetics.
+- Added new component: `setFitnessGoal.vue`.
+- Made all cards in the dashboard clickable, routing to their respective "edit" routes.
+- Added components for editing:
+  - `EditActivity.vue`
+  - `EditWorkout.vue`
+  - `EditGoal.vue`
+
 ## Jul 29, 2024
 - Update Dashboard: Display Workout Routines, Activities Logged.